ddprocess produces html output

+++ bin/ddproces.sh
@@ -1,4 +1,5 @@
 #! /bin/bash
+CMD_THIS_HTML="<code>$(basename $0)</code> <var>${@}</var>"
 
 NAME=ddprocess
 VERSION="0.90"
@@ -246,7 +247,13 @@ save_n_update () {
           --host-ip="147.102.110.69" \
           --host-dir="${SOL_DIR}/${YEAR}/${DOY_3C}/" \
           --product-filename="${trg_f}.Z" ; then
-  echo "Final ${3} saved/recorded: ${trg_f}.Z"
+#  printf "[SAVED::%s] File %s saved as %s.Z; " \
+#          "$1" "$src_f" "$trg_f"
+#  printf "Database entry inserted at 147.102.110.69%s\n" \
+#          "${SOL_DIR}/${YEAR}/${DOY_3C}/"
+  printf "<tr><td>%s</td><td><code>%s</code></td><td><code>%s</code></td><td><code>%s</code></td><td>%s</td><td>%s</td></tr>\n" \
+          "${3}" "${src_f}" "${trg_f}.Z" "${SOL_DIR}/${YEAR}/${DOY_3C}/" \
+          "${START_OF_DAY_STR/ /_}" "${END_OF_DAY_STR/ /_}"
   return 0
 else
   echoerr "ERROR. Failed to save/record ${1} sinex : ${trg_f}"
@@ -301,6 +308,14 @@ TIME_STAMP_FILE=".ddprocess-ts-${BASHPID}"
 touch ${TIME_STAMP_FILE}
 date > ${TIME_STAMP_FILE}
 
+## ////////////////////////////////////////////////////////////////////////////
+##  REPORT ..
+## ////////////////////////////////////////////////////////////////////////////
+#echo "[MAIN::] ${NAME} ${VERSION}-${REVISION} ${LAST_UPDATE}"
+printf "<html>\n"
+printf "<head><title>DDProcess REPORT</title></head>\n"
+printf "<h1>General Information</h1>\n"
+printf "<table><head></head><body>\n"
 ## ////////////////////////////////////////////////////////////////////////////
 ## GET/EXPAND COMMAND LINE ARGUMENTS
 ## ////////////////////////////////////////////////////////////////////////////
@@ -436,6 +451,16 @@ do
   shift
 done
 
+printf "<tr><td>Campaign</td><td><var>%s</var></td></tr>\n" \
+        "${CAMPAIGN}"
+printf "<tr><td>User Host</td><td><var>%s@%s</var></td></tr>\n" \
+        "${USER}" "${HOSTNAME}"
+printf "<tr><td>Command</td><td>%s</td></tr>\n" \
+        "$CMD_THIS_HTML"
+printf "<tr><td>ddprocess</td><td><var>%s<var></td></tr>\n" \
+        "${NAME} ${VERSION}-${REVISION} ${LAST_UPDATE}"
+printf "</body></table>\n"
+
 ## ////////////////////////////////////////////////////////////////////////////
 ##  LOGFILE, STDERR & STDOUT
 ##  ---------------------------------------------------------------------------
@@ -566,6 +591,14 @@ if ! END_OF_DAY_STR=$(ydoy2dt "${YEAR}" "${DOY_3C}" "23" "59" "30") ; then
 fi
 
 ## ////////////////////////////////////////////////////////////////////////////
+##  REPORT ..
+## ////////////////////////////////////////////////////////////////////////////
+#printf "[DATEP::] Processing date: %s from %s to %s.\n" \
+#      "${YEAR}-${DOY_3C}" \
+#      "${START_OF_DAY_STR}" \
+#      "${END_OF_DAY_STR}"
+
+## ////////////////////////////////////////////////////////////////////////////
 ##  DOWNLOAD RINEX FILES
 ##  ---------------------------------------------------------------------------
 ##  Use the program rnxdwnl.py to download all available rinex files for
@@ -580,7 +613,9 @@ fi
 ##  Lastly, copy and uncompress (.Z && crx2rnx) the files in the campaign
 ##+ directory /RAW.
 ## ////////////////////////////////////////////////////////////////////////////
-if test 1 -eq 2 ; then
+if test 2 -eq 1 ; then
+
+printf "<h2>Rinex</h2>"
 >.rnxsta.dat ## temporary file
 
 ##  download the rinex files for the input network; the database knows 
@@ -660,8 +695,13 @@ done
 >.station-names.dat
 for sta in "${STA_ARRAY[@]}"; do echo $sta >> .station-names.dat; done
 
-echo "Number of stations available: ${#STA_ARRAY[@]}/${MAX_NET_STA}"
-echo "Number of reference stations: ${#REF_STA_ARRAY[@]}"
+## ////////////////////////////////////////////////////////////////////////////
+##  REPORT ..
+## ////////////////////////////////////////////////////////////////////////////
+printf "<p>Number of stations available: %s/%s</p>\n" \
+        "${#STA_ARRAY[@]}" "${MAX_NET_STA}"
+printf "<p>Number of reference stations: %s</p>\n" \
+        "${#REF_STA_ARRAY[@]}"
 fi
 ## ////////////////////////////////////////////////////////////////////////////
 ##  DOWNLOAD IONOSPHERIC MODEL FILE
@@ -670,6 +710,8 @@ fi
 ##  TODO :: mysql ...
 ##  If such a file does not exist, download CODE's ionospheric file.
 ## ////////////////////////////////////////////////////////////////////////////
+printf "<h2>Products</h2>"
+
 ION_DOWNLOADED=0
 
 if mysql -h "${DB_HOST}" \
@@ -692,7 +734,7 @@ if mysql -h "${DB_HOST}" \
       | awk '{print $1$2}' > .procsta-answer.dat \
     && grep "[a-z,A-Z]*" .procsta-answer.dat &>/dev/null ; then
   my_ion_file=$(cat .procsta-answer.dat)
-  echoerr "ERROR. NTUA's ION filo found but ddprocess does not yet handle that!"
+  echoerr "ERROR. NTUA's ION file found but ddprocess does not yet handle that!"
   echoerr "Write more code bitch!"
 else
   echoerr "[WARNING]. No NTUA .ION file found."
@@ -709,7 +751,7 @@ fi
 ##+ the following will download the best possible products; for more info, see
 ##+ the bernutils module documentation.
 ## ////////////////////////////////////////////////////////////////////////////
-if test 2 -eq 1 ;  then
+if test 1 -eq 1 ;  then
 ##  will we need an ion file ?
 if test ${ION_DOWNLOADED} -eq 1; then
   if ! handle_dd_products.py \
@@ -718,7 +760,8 @@ if test ${ION_DOWNLOADED} -eq 1; then
           --analysis-center="${AC}" \
           --datapool="${D}" \
           --destination="${P}/${CAMPAIGN}/ORB" \
-          --satellite-system="${SAT_SYS}" ; then
+          --satellite-system="${SAT_SYS}" \
+          --report=html; then
     echoerr "ERROR. Failed to download/copy/uncompress products."
     exit 1
   fi
@@ -730,7 +773,8 @@ else
           --datapool="${D}" \
           --destination="${P}/${CAMPAIGN}/ORB" \
           --satellite-system="${SAT_SYS}" \
-          --download-ion ; then
+          --download-ion \
+          --report=html; then
     echoerr "ERROR. Failed to download/copy/uncompress products."
     exit 1
   fi
@@ -745,7 +789,7 @@ fi
 ##+ first 6 hours of the next day (a bernese thing). We need to merge all
 ##+ files to a final one.
 ## ////////////////////////////////////////////////////////////////////////////
-if test 2 -eq 1 ; then
+if test 1 -eq 1 ; then
 ## temporary file to hold getvmf1.py output
 TMP_FL=.vmf1-${YEAR}${DOY}.dat
 
@@ -766,6 +810,20 @@ else
     cat ${fl} >> ${MERGED_VMF_FILE}
   done
 fi
+
+##  write a nice report like the one given by handle_products.
+if ! mapfile -t VMF_FL_ARRAY < <(cat ${TMP_FL} | awk '{print $2}') ; then
+  echoerr "ERROR. Failed to report VMF1 grid file(s). Strange !!"
+  exit 1
+fi
+let c=${#VMF_FL_ARRAY[@]}-1
+tmp_="<code>${VMF_FL_ARRAY[0]}"
+for i in `seq 1 $c`; do tmp_="${tmp_}, ${VMF_FL_ARRAY[${i}]}"; done
+tmp_="${tmp_}</code>"
+printf "<p>Product type: <strong>vmf1</strong> :\
+  downloaded file(s) %s ; moved to %s</p>\n" \
+      "$tmp_" "$MERGED_VMF_FILE"
+
 rm ${TMP_FL} ## remove temporary file
 fi
 ## ////////////////////////////////////////////////////////////////////////////
@@ -814,7 +872,7 @@ awk -v num_of_clu=${STATIONS_PER_CLUSTER} -f \
 ##+      PRELIM_SOLUTION_ID  = 'FFP'
 ##+      REDUCED_SOLUTION_ID = 'FFR1'
 ## ////////////////////////////////////////////////////////////////////////////
-
+printf "<h2>Solution Identifiers</h2>"
 ##  Final (ambiguity-fixed) results
 FINAL_SOLUTION_ID="${SOLUTION_ID}"
 
@@ -836,11 +894,16 @@ else
   REDUCED_SOLUTION_ID="${SOLUTION_ID%?}R"
 fi
 
+##  report ..
+printf "<p><strong>Final Solution</strong>       : <var>%s</var><br>\n" "$FINAL_SOLUTION_ID"
+printf "<strong>Size-Reduced Solution</strong>: <var>%s</var><br>\n" "$REDUCED_SOLUTION_ID"
+printf "<strong>Preliminery Solution</strong> : <var>%s</var></p>\n" "$PRELIM_SOLUTION_ID"
+
 ## ////////////////////////////////////////////////////////////////////////////
 ##  SET VARIABLES IN THE PCF FILE
 ##  ---------------------------------------------------------------------------
 ## ////////////////////////////////////////////////////////////////////////////
-
+printf "<h2>Process Control File (PCF)</h2>"
 ##  Bernese has no 'MIXED' satellite system; this defaults to 'GPS/GLO'.
 if test "${SAT_SYS}" == "MIXED"; then
   BERN_SAT_SYS="GPS/GLO"
@@ -932,7 +995,8 @@ fi
 ##  COPY PRODUCTS TO HOST; UPDATE DATABASE ENTRIES
 ##  ---------------------------------------------------------------------------
 ## ////////////////////////////////////////////////////////////////////////////
-if test 1 -eq 2 ; then
+printf "<h2>Saved Products</h2>"
+if test 1 -eq 1 ; then
 ##  warning: in the db mixed := GPS+GLO
 if test "${SAT_SYS^^}" = "MIXED"; then
   DB_SAT_SYS="GPS+GLO"
@@ -956,6 +1020,14 @@ fi
 ##  argv2 -> campaign dir (e.g. 'SOL')
 ##  argv3 -> product type (e.g. 'SINEX')
 
+printf "<table style=\"width:100%%\" id=\"t02\" border=\"1\">"
+printf "<thead><th>Type</th> \
+  <th>Source File</th> \
+  <th>Target File</th> \
+  <th>Path</th> \
+  <th>From</th>\
+  <th>To</th></thead>\n<tbody>"
+
 ##  final tropospheric sinex
 if ! save_n_update TRO ATM TRO_SNX ; then exit 1 ; fi
 
@@ -970,6 +1042,10 @@ if ! save_n_update NQ0 SOL NQ R ; then exit 1 ; fi
 
 ## final coordinates
 if ! save_n_update CRD STA CRD_FILE ; then exit 1 ; fi
+
+printf "</tbody>\n"
+printf "<caption>List of saved products.</caption>\n"
+printf "</table>"
 fi
 ## ////////////////////////////////////////////////////////////////////////////
 ##  COMPILE (NON-FATAL) ERROR/WARNINGS FILE
@@ -982,17 +1058,44 @@ if ! >${WRN_FILE} ; then
 fi
 
 find ${P}/${CAMPAIGN}/OUT/WRN${DOY_3C}0*.SUM -not -empty -ls -exec \
-      cat {} 1>>${WRN_FILE} \; ## match files e.g 'WRN0010003.SUM'
+      cat {} 1>>${WRN_FILE} 2>/dev/null \; ## match files e.g 'WRN0010003.SUM'
 
 find ${P}/${CAMPAIGN}/OUT/*${YEAR}${DOY_3C}0.ERR -not -empty -ls -exec \
-      cat {} 1>>${WRN_FILE} \; ## match files e.g 'RNX150010.ERR'
+      cat {} 1>>${WRN_FILE} 2>/dev/null \; ## match files e.g 'RNX150010.ERR'
 
 echo "Warnings file created as ${WRN_FILE}"
+
+## ////////////////////////////////////////////////////////////////////////////
+##  ADDNEQ SUMMARY TO HTML
+##  ---------------------------------------------------------------------------
+## ////////////////////////////////////////////////////////////////////////////
+printf "<h2>Ambiguity Resolution</h2>"
+##  the ambiguity summary file
+AMBSM=${P}/${CAMPAIGN}/OUT/AMB${YEAR:2:2}${DOY_3C}0.SUM
+python - <<END
+import sys, bernutils.bamb
+try:
+  ambf = bernutils.bamb.AmbFile( "${AMBSM}" )
+  ambf.toHtml()
+except:
+  print>>sys.stderr,'ERROR. Cannot translate amb file to html!'
+  sys.exit(1)
+sys.exit(0)
+END
+if test "$?" -ne 0 ; then exit 1 ; fi
+
+printf "<h2>ADDNEQ Summary</h2>"
+##  the final ADDNEQ summary file should be
+ADNQ=${P}/${CAMPAIGN}/OUT/${FINAL_SOLUTION_ID}${YEAR:2:2}${DOY_3C}0.OUT
+adnq2html.py --addneq-file="${ADNQ}" \
+          --table-entries='latcor,loncor,hgtcor,dn,de,du,adj' \
+          --warnings-str='dn=.01,de=.01,du=.01'
 
 ## ////////////////////////////////////////////////////////////////////////////
 ##  REMOVE CAMPAIGN FILES
 ##  ---------------------------------------------------------------------------
 ## ////////////////////////////////////////////////////////////////////////////
+if test 1 -eq 2 ; then
 ##  we are going to remove any file in the campaign-specific folders, newer
 ##+ than .ddprocess-time-stamp (i.e. the stamp file), except from symlinks.
 find -P ${P}/${CAMPAIGN}/ \
@@ -1000,5 +1103,6 @@ find -P ${P}/${CAMPAIGN}/ \
       -type f \
       -newer ${TIME_STAMP_FILE} \
       ! name ${WRN_FILE}
+fi
 
 exit 0
